+++ linearmodels/panel/results.py
@@ -746,7 +746,7 @@ def compare(results, precision='tstats'):
         Set of results to compare.  If a dict, the keys will be used as model
         names.  An OrderedDict will preserve the model order the comparisons.
     precision : {'tstats','std_errors', 'std-errors', 'pvalues'}
-        Estimator precision estimator to include in the comparrison output.
+        Estimator precision estimator to include in the comparison output.
         Default is 'tstats'.
 
     Returns
@@ -766,7 +766,7 @@ class PanelModelComparison(_ModelComparison):
         Set of results to compare.  If a dict, the keys will be used as model
         names.  An OrderedDict will preserve the model order the comparisons.
     precision : {'tstats','std_errors', 'std-errors', 'pvalues'}
-        Estimator precision estimator to include in the comparrison output.
+        Estimator precision estimator to include in the comparison output.
         Default is 'tstats'.
     """
     _supported = (PanelEffectsResults, PanelResults, RandomEffectsResults)